+++ lua/terrortown/lang/fr.lua
@@ -2682,3 +2682,6 @@ L.vehicle_enter = "Appuyez sur [{usekey}] pour entrer dans le véhicule"
 --</ol>
 --
 --]]
+
+-- 2025-01-19
+--L.help_rolelayering_enable = "The red and green border around the icon shows if the role is currently enabled. Right click on an icon to quickly enable/disable that role."