+++ CHANGELOG.md
@@ -11,6 +11,7 @@ All notable changes to TTT2 will be documented here. Inspired by [keep a changel
   - This also includes a Garry's Mod version check which triggers a warning if TTT2 is not compatible. First baseline version is '240313' (by @NickCloudAT)
 - Added `GM:TTT2PlayDeathScream` hook to cancel or overwrite/change the deathscream sound that plays, when you die (by @NickCloudAT)
 - Added support for "toggle_zoom" binds to trigger the radio commands menu (by @TW1STaL1CKY)
+- Added option to use right click to enable/disable roles in the role layering menu (by @TimGoll)
 - Added a menu to allow admins to inspect, in detail, how and why roles are distributed as they are (by @nike4613)
 
 ### Fixed
@@ -56,6 +57,7 @@ All notable changes to TTT2 will be documented here. Inspired by [keep a changel
   - Modes: 0 = No session limit, 1 = Default TTT, 2 = Only time limit, 3 = Only round limit
 - Moved all role-related admin options into the "Roles" menu (by @nike4613)
 - Improved description of role layering (by @nike4613)
+- Improved the role layering menu by showing which role is enabled and which is disabled (by @TimGoll)
 
 ## [v0.14.0b](https://github.com/TTT-2/TTT2/tree/v0.14.0b) (2024-09-20)